Create Use Cases Prototyping
If you have any suggestions/questions feel free to post them on Slack in General. 
It is VERY important that Front-End and Back-End cooperate to help make sure the site we are both trying to build is the same.
To that end, we both need to work together to decide the Use Cases. Things can be changed obvi, but it would be nice to be
99% sure of things before we start coding.